+++ CHANGELOG.md
@@ -10,6 +10,9 @@ The `Unreleased` section name is replaced by the expected version of next releas
 
 ### Added
 ### Changed
+
+- Targeted `Equinox`.* v `2.0.0-rc3`
+
 ### Removed
 ### Fixed